refactor(secrets): move bare env-var marker helper to model-auth-markers
Move isBareEnvVarNameApiKeyMarker to model-auth-markers.ts so audit.ts
stays under the 700 effective-line lint limit. The helper now resolves
the provider's canonical config SecretRef internally, keeping the audit
call site to a single expression.

No behavior change.
